% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chivalry}
\alias{chivalry}
\title{Chivalry ratings of Gryffindor's and Slytherin's}
\format{
A data frame with 80 rows and 3 variables.

\describe{
  \item{\code{id}}{Participant's ID}
  \item{\code{house}}{Participant's house}
  \item{\code{chivalry}}{Participant's chivalry rating}
}
}
\source{
Fictional dataset.
}
\usage{
chivalry
}
\description{
A dataset containing chivalry ratings for 80 members of the Hogwarts houses
Gryffindor and Slytherin. This was used as an example for an independent
t-Test.
}
\keyword{datasets}
