% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_g.R
\name{compute_hedges_g}
\alias{compute_hedges_g}
\title{Combine Effect Sizes in Single Function}
\usage{
compute_hedges_g(
  data,
  id = NULL,
  grouping_factor,
  variable,
  paired = FALSE,
  detailed = FALSE
)
}
\arguments{
\item{data}{Data frame containing the data}

\item{id}{Subject identifier}

\item{grouping_factor}{Factor, used for the comparisons between groups}

\item{variable}{Numeric, the dependent variable}

\item{paired}{Indicator, if repeated measures should be used.}

\item{detailed}{Indicator, if detailed information (such as means, standard
deviations, etc.) should be returned. Default ist \code{FALSE}}
}
\value{
A tibble
}
\description{
This function computes pairwise independent effect sizes (Hedges' \eqn{g} and
\eqn{g*}) for a continuous variable based on a grouping variable.
}
