% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effsizes.R
\name{effsize_repeated}
\alias{effsize_repeated}
\title{Compute Effect Sizes for Repeated Measures}
\usage{
effsize_repeated(
  data,
  id,
  grouping_factor,
  grouping_levels,
  variable,
  detailed = FALSE
)
}
\arguments{
\item{data}{Data frame containing the data}

\item{id}{Subject identifier}

\item{grouping_factor}{Factor, used for the comparisons between groups}

\item{grouping_levels}{Concatenated string, the levels of the grouping factor
that should be compared}

\item{variable}{Numeric, the dependent variable}

\item{detailed}{Indicator, if detailed information (such as means, standard
deviations, etc.) should be returned. Default ist \code{FALSE}}
}
\value{
A tibble
}
\description{
Compute Effect Sizes for Repeated Measures
}
