% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam_stress}
\alias{exam_stress}
\title{Exam Stress}
\format{
A data frame containing 288 rows and 4 variables.
\describe{
    \item{\code{id}}{Participant's ID}
    \item{\code{preparation}}{Factor indicating whether a student prepared for the exam or not}
    \item{\code{time}}{Factor indicating time of stress measurement}
    \item{\code{stress}}{Observed stress level on a scale vom 0 (lowest) to 50 (highest)}
}
}
\source{
Fictional dataset
}
\usage{
exam_stress
}
\description{
A dataset containing stress levels for 96 participants before, during and
after an exam, split by whether they prepared or did not prepare for this
exam. This was used as an example for a mixed ANOVA.
}
\keyword{datasets}
