% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effsize.R
\name{get_effsize}
\alias{get_effsize}
\title{Compute Effect Sizes}
\usage{
get_effsize(
  data,
  grouping_factor,
  grouping_levels,
  variable,
  paired = FALSE,
  participant_id,
  detailed = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{grouping_factor}{Factor, a variable used to group upon}

\item{grouping_levels}{Strings, two concatenated strings to use as factor
levels for comparison}

\item{variable}{Numeric, a dependent variable}

\item{paired}{Indicator, if \code{TRUE}, repeated measures effect sizes will
be calculated}

\item{participant_id}{A column in the data frame containing a participant
identifier. This is needed only for repeated measures effect sizes}

\item{detailed}{Indicator, if \code{FALSE} (default) detailed statistics will
be omitted}
}
\value{
A tibble
}
\description{
This function computes effect sizes for continuous variables based on
grouping variables.
}
