% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ice_cream}
\alias{ice_cream}
\title{Yumminessof different ice cream flavors}
\format{
A data frame with 295 rows and 4 variables.

\describe{
  \item{id}{Participant's ID}
  \item{flavor}{Ice cream flavor}
  \item{sex}{Participant's sex}
  \item{yumminess}{Ice cream's yumminess}
}
}
\source{
Fictional dataset.
}
\usage{
ice_cream
}
\description{
A dataset containing yumminess ratings of 295 voluntary tasters for different
ice cream flavors. This was used as an example dataset for a factorial ANOVA.
}
\keyword{datasets}
