% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patience}
\alias{patience}
\title{Patience ratings of Hufflepuffs}
\format{
A data frame with 106 rows and 3 variables.

\describe{
  \item{id}{Participant's ID}
  \item{measurement}{Time of measurement in relation to the school year}
  \item{patience}{Participant's patience}
}
}
\source{
Fictional dataset.
}
\usage{
patience
}
\description{
A dataset containing patience ratings for 106 members of the Hogwarts
houseHufflepuff. This was used as an example for a dependent t-Test.
}
\keyword{datasets}
