% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phantasialand}
\alias{phantasialand}
\title{Joy of Phantasialand's visitors}
\format{
A data frame with 186 rows and 3 variables.

\describe{
  \item{id}{Participant's ID}
  \item{backpack}{Participant's backpack}
  \item{joy}{Participant's joy}
}
}
\source{
Fictional dataset.
}
\usage{
phantasialand
}
\description{
A dataset containing joy ratings for 186 visitors of the Phantasialand. This
was used as an example dataset for a one-way ANOVA.
}
\keyword{datasets}
