% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{placebo_amplification}
\alias{placebo_amplification}
\title{Placebo Amplification Data}
\format{
A data frame with 172 rows and 13 variables.
\describe{
    \item{\code{id}}{Participant's age}
    \item{\code{age}}{Age}
    \item{\code{sex}}{Sex, a factor with two levels ("Female" or "Male")}
    \item{\code{treatment}}{The randomly assigned treatment, a factor with two levels ("TAU" = treatment as usual or "EG" = experimental group)}
    \item{\code{measurement}}{Number of measurement}
    \item{\code{exp_size}}{Participant's expected size of the improvement due to the antidepressant}
    \item{\code{exp_certain}}{Participant's expected certainty of an improvement due to the antidepressant}
    \item{\code{exp_speed}}{Participant's expected speed of the improvement due to the antidepressant}
    \item{\code{cgi_s}}{Clinical Global Impressions, severity score}
    \item{\code{bdi}}{Beck's Depression Inventory score}
    \item{\code{shaps}}{Snaith-Hamilton Pleasure Scale score}
    \item{\code{who}}{WHO Well-Being Questionnaire score}
    \item{\code{hamd}}{Hamilton Rating Scale for Depression score}
}
}
\source{
\href{https://www.karger.com/Article/Abstract/505855}{Claus, B. B., Scherbaum, N., & Bonnet, U. (2020).
Effectiveness of an Adjunctive Psychotherapeutic Intervention Developed for Enhancing the Placebo Effect of
Antidepressants Used within an Inpatient-Treatment Program of Major Depression: A Pragmatic Parallel-Group,
Randomized Controlled Trial. Psychotherapy and Psychosomatics, 1-3. https://doi.org/10.1159/000505855}
}
\usage{
placebo_amplification
}
\description{
A dataset containing data from a randomized clinical trial sought to amplify
the placebo effect of antidepressants for 43 participants. This was used as a
real world example for repeated measures corrlation.
}
\keyword{datasets}
