% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ravenclaw}
\alias{ravenclaw}
\title{Intelligence ratings of Ravenclaw's}
\format{
A data frame with 49 rows and 2 variables.

\describe{
    \item{\code{id}}{Participant's ID}
    \item{\code{intelligence}}{Participant's
  intelligence}
}
}
\source{
Fictional dataset.
}
\usage{
ravenclaw
}
\description{
A dataset containing intelligence ratings for 48 members of the Hogwarts
house of Ravenclaw. This was used as an example for a one-sample t-Test.
}
\keyword{datasets}
