% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reproducibility_project}
\alias{reproducibility_project}
\title{Reproducibility Poject Psychology}
\format{
A data frame with 168 rows and 15 variables.

\describe{
    \item{\code{study_title_o}}{Original study's title}
    \item{\code{journal_o}}{Journal}
    \item{\code{first_author}}{First author}
    \item{\code{senior_author_o}}{Senior author}
    \item{\code{institution_prestige_1st_author_o}}{First author's institution prestige}
    \item{\code{institution_prestige_senior_author_o}}{Senior author's institution prestige}
    \item{\code{surprising_result_o}}{Surprising result? (see original codebook for more details)}
    \item{\code{exciting_result_o}}{Exciting result? (see original codebook for more details)}
    \item{\code{pval_original}}{p-value in original study}
    \item{\code{pval_replication}}{p-value in replication study}
    \item{\code{es_original}}{Effect size (r) in original study}
    \item{\code{es_replication}}{Effect size (r) in replication study}
    \item{\code{power_r}}{Power in replication study}
    \item{\code{sig_original}}{Indicator, if p was significant in original study (Note: three studies had a p between .05 and .06. Those were countas as significant as well)}
    \item{\code{sig_replication}}{Indicator, if p was significant in replication study}
}
}
\source{
\href{https://osf.io/ezcuj}{Open Science Framework repository of the original study.}
}
\usage{
reproducibility_project
}
\description{
A dataset containing a reduced version of the original published dataset for
the ingenious project by the Open Science Collaboration.
}
\keyword{datasets}
