% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runners}
\alias{runners}
\title{Perceived endurance of runners}
\format{
A data frame with 153 rows and 3 variables.

\describe{
  \item{id}{Participant's ID}
  \item{music_type}{Type of music listened to}
  \item{endurance}{Participant's perceived endurance}
}
}
\source{
Fictional dataset.
}
\usage{
runners
}
\description{
A dataset containing endurance ratings of 51 voluntary runners, running while
listening to different types of music. This was used as an example dataset
for a one-way repeated measures ANOVA.
}
\keyword{datasets}
