% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runners_motivation}
\alias{runners_motivation}
\title{Perceived endurance of runners}
\format{
A data frame with 153 rows and 4 variables.

\describe{
  \item{\code{id}}{Participant's ID}
  \item{\code{music_type}}{Type of music listened to}
  \item{\code{motivation}}{Did the runner have motivation for that run?}
  \item{\code{endurance}}{Participant's perceived endurance}
}
}
\source{
Fictional dataset.
}
\usage{
runners_motivation
}
\description{
A dataset containing endurance ratings of 51 voluntary runners, running while
listening to different types of music while motivated or not. This was used
as an example dataset for a two-way repeated measures ANOVA.
}
\keyword{datasets}
