% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typing}
\alias{typing}
\title{Speed and Accuracy Data}
\format{
A data frame with 435 rows and 4 variables

\describe{
    \item{id}{Participant's ID}
    \item{trial}{Trial}
    \item{speed}{Participant's typing speed in keystrokes per minute}
    \item{accuracy}{Participant's accuracy in percent with 100 indicating that no mistakes were made}
}
}
\source{
Fictional dataset
}
\usage{
typing
}
\description{
A dataset containing speed and accuracy data for 87 participants trying to
type as fast as they can. This was used as an example for repeated measures
correlation (rmcorr)
}
\keyword{datasets}
