% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_park}
\alias{water_park}
\title{Water Park Data}
\format{
A data frame with 100 rows and 5 variables.
\describe{
    \item{\code{day_id}}{A uniqe identifier of observed days}
    \item{\code{temperature}}{Day's temperature}
    \item{\code{swimmers}}{Day's swimmers/guests}
    \item{\code{sales}}{Day's sales}
    \item{\code{beatings}}{Observed beatings}
}
}
\source{
Fictional dataset
}
\usage{
water_park
}
\description{
A dataset containing different dependent variables measured on 100 different
days in a random water park. This was used as an example for correlation
analyses.
}
\keyword{datasets}
