% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effsize.R
\name{compute_effsize}
\alias{compute_effsize}
\title{Compute Pairwise Effect Sizes}
\usage{
compute_effsize(
  data,
  grouping_factor,
  variable,
  paired = FALSE,
  participant_id,
  detailed = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{grouping_factor}{Factor, a variable used to group upon}

\item{variable}{Numeric, a dependent variable}

\item{paired}{Indicator, if \code{TRUE}, repeated measures effect sizes will
be calculated}

\item{participant_id}{A column in the data frame containing a participant
identifier. This is needed only for repeated measures effect sizes}

\item{detailed}{Indicator, if \code{FALSE} (default) detailed statistics will
be omitted}
}
\value{
A tibble
}
\description{
This function computes pairwise independent effect sizes (Hedges' \eqn{g} and
\eqn{g*}) for a continous variable based on a grouping variable.
}
