% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effsize.R
\name{get_effsize}
\alias{get_effsize}
\title{Compute Effect Sizes}
\usage{
get_effsize(data, grouping_factor, grouping_levels, variable, detailed = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{grouping_factor}{Factor, a variable used to group upon}

\item{grouping_levels}{Strings, two concatenated strings to use as factor
levels for comparison}

\item{variable}{Numeric, a dependent variable}

\item{detailed}{Indicator, if \code{FALSE} (default) detailed statistics
will be omitted}
}
\value{
A tibble
}
\description{
This function computes effect sizes for continuous variables based on
grouping variables.
}
